Introduce interface SmofUpdate
